requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.4.1
